requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
dnspython==2.7.0
Events==0.5
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.9.0
motor==3.7.1
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
openai==1.78.1
opensearch-py==2.8.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tqdm==4.67.1
twilio==9.6.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
